feat: implement heat system parity with HeatSources and IHeatSource API

+++ src/main/java/com/chouzz/skyresourcereforge/api/IHeatSource.java
@@ -4,5 +4,5 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 
 public interface IHeatSource {
-    float getHeatValue();
+    int getHeatValue();
 }
