+++ src/main/java/com/chouzz/skyresourcereforge/block/entity/CasingBlockEntity.java
@@ -9,18 +9,18 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
 public class CasingBlockEntity extends BlockEntity implements IHeatSource {
-    private float currentHeat = 0;
+    private int currentHeat = 0;
 
     public CasingBlockEntity(BlockPos pos, BlockState state) {
         super(ModBlockEntities.CASING.get(), pos, state);
     }
 
     @Override
-    public float getHeatValue() {
+    public int getHeatValue() {
         return currentHeat;
     }
 
-    public void setHeatValue(float heat) {
+    public void setHeatValue(int heat) {
         this.currentHeat = heat;
         setChanged();
     }
@@ -28,12 +28,12 @@ public class CasingBlockEntity extends BlockEntity implements IHeatSource {
     @Override
     protected void saveAdditional(CompoundTag tag, HolderLookup.Provider registries) {
         super.saveAdditional(tag, registries);
-        tag.putFloat("heat", currentHeat);
+        tag.putInt("heat", currentHeat);
     }
 
     @Override
     protected void loadAdditional(CompoundTag tag, HolderLookup.Provider registries) {
         super.loadAdditional(tag, registries);
-        currentHeat = tag.getFloat("heat");
+        currentHeat = tag.getInt("heat");
     }
 }
